commit amilkar

+++ Entidades/Pagos y Facturas/Pago.cs
@@ -31,7 +31,7 @@ namespace Entidades.Pagos_y_Facturas
     public class Pago
     {
         public int Id { get; set; }
-        public decimal ValorIngresado { get; set; }
+        public double ValorIngresado { get; set; }
         public DateTime FechaPago { get; set; }
         public int IdInscripcion { get; set; }
